add breakout_game
The objective of the game is to use the paddle to bounce the ball towards the bricks and destroy them. You win when all the bricks are destroyed. But if the ball falls below the paddle into the fiery pits of doom, you lose the game. The paddle has "zones" which bounce the ball in different angles. Try out the game and you will notice it. If the ball bounces at the center, it goes straight up. As it goes to the sides, it's angle and speed increase. This gives you a certain degree of control over the direction of the ball.
#573